Start to refactor backend funding spaces get
