+++ src/routes/fundingSpaces.ts
@@ -1,6 +1,7 @@
 import express from 'express';
 import { passAsyncError } from '../middleware/error/passAsyncError';
 import * as controller from '../controllers/fundingSpaces';
+import { parseQueryString } from '../utils/parseQueryString';
 
 export const fundingSpacesRouter = express.Router();
 
@@ -11,8 +12,15 @@ fundingSpacesRouter.get(
     const organizationIds = !organizationId
       ? undefined
       : ((Array.isArray(organizationId)
-          ? organizationId
-          : [organizationId]) as string[]);
+        ? organizationId
+        : [organizationId]) as string[]);
+
+    const fundingMap = parseQueryString(req, 'fundingMap');
+    if (fundingMap === 'true') {
+      const fundingSpacesMap = await controller.getFundingSpaceMap();
+      res.send({ fundingSpacesMap });
+    }
+
     const fundingSpaces = await controller.getFundingSpaces(
       req.user,
       organizationIds
